refactor(blorktools): image2texture prerender
