+++ packages/blorktools/src/asset_debugger/modals/html-editor-modal/html-editor-modal.js
@@ -37,7 +37,7 @@ import {
 import { updateHtmlIcons } from '../../panels/mesh-panel/mesh-panel';
 // Import Three.js the same way as other files in the codebase
 import * as THREE from 'three';
-import { getIsPreviewActive, setLastTextureUpdateTime } from '../../util/custom-animation/animation-util';
+import { isPreviewActive, setLastTextureUpdateTime } from '../../util/custom-animation/animation-util';
 import { previewHtml } from '../../util/custom-animation/preview-util';
 import { cleanupThreeJsPreview, frameBuffer, previewRenderTarget } from '../../util/custom-animation/threejs-util.js';
 import { getHtmlSettingsForMesh, loadHtmlForMesh, loadSettingsForMesh, saveHtmlForMesh, saveSettingsForMesh } from '../../util/mesh-data-util.js';
@@ -412,7 +412,7 @@ export function initHtmlEditorModal() {
                 saveSettingsForMesh(meshId, settings);
                 
                 // Update CSS3D preview if active
-                if (getIsPreviewActive()) {
+                if (isPreviewActive) {
                     try {
                         // For CSS3D preview - update animation speed via CSS
                         const css3dIframe = document.getElementById('css3d-panel-iframe');
@@ -457,7 +457,7 @@ export function initHtmlEditorModal() {
             window.showPreviewBorders = showWireframeCheckbox.checked;
             
             // If preview is active, update it immediately
-            if (getIsPreviewActive() && previewRenderTarget) {
+            if (isPreviewActive && previewRenderTarget) {
                 // Force a texture update
                 setLastTextureUpdateTime(0);
                 showStatus(`Borders ${showWireframeCheckbox.checked ? 'enabled' : 'disabled'}`, 'info');
