+++ packages/blorktools/src/asset_debugger/util/custom-animation/animation-util.js
@@ -5,8 +5,43 @@ import { createTextureFromIframe, injectUnifiedAnimationDetectionScript } from '
 
 const ANALYSIS_DURATION_MS = 30000; // 30 seconds - matches preRenderMaxDuration
 
+export let finalProgressAnimation = false;
+export let finalProgressStartTime = 0;
+export let finalProgressDuration = 800; // Duration of final progress animation in ms
+
+let animationChangeThreshold = 0.008; // Lower threshold for detecting change (was 0.01)
+let animationIdleThreshold = 30; // Number of similar frames before considering animation ended
+let frameChangeRates = []; // Store recent frame change rates
+let animationIdleCount = 0;
+let previousChangeFrequency = 0; // Track previous change frequency
+let preRenderStartTime = 0;
+export let preRenderMaxDuration = 30000; // Increased to 30 seconds for longer animations
+// Image2texture variables
+export let animationDetectionSensitivity = 0.85; // Increased from 0.5 to 0.85 for much stricter detection by default
+let animationEndTime = 0;
+let animationStartTime = 0;
+export let animationStartDetected = false;
+export let animationDetected = false;
+export let preRenderedFrames = [];
+export let animationDuration = 0; // Store detected animation duration
+export let preRenderingInProgress = false;
+export let isAnimationFinite = false;
+
+// Update references to use internal functions instead of imports
+// These variables will be exported and should be used by preview-util.js
+export let isPreviewActive = true; // Exported for use in preview-util.js
+export let isPreviewAnimationPaused = false; // Exported for use in preview-util.js
+export let lastTextureUpdateTime = 0; // Exported for use in preview-util.js and threejs-util.js
+
+// TIMING STATE - This module owns all animation timing
+let animationPlaybackStartTime = 0; // When playback actually started
+let animationCaptureStartTime = 0;  // When capture started (for offset calculations)
+// Active playback values
+let playbackStartTime = 0;
+let isPlaybackActive = false;
+
 // Debug reporting function for animation analysis
-function logAnimationAnalysisReport(renderType, data) {
+export function logAnimationAnalysisReport(renderType, data) {
     const {
         frameCount,
         duration,
@@ -34,41 +69,6 @@ function logAnimationAnalysisReport(renderType, data) {
         metrics
     });
 }
-
-let finalProgressAnimation = false;
-let finalProgressStartTime = 0;
-let finalProgressDuration = 800; // Duration of final progress animation in ms
-
-let animationChangeThreshold = 0.008; // Lower threshold for detecting change (was 0.01)
-let animationIdleThreshold = 30; // Number of similar frames before considering animation ended
-let frameChangeRates = []; // Store recent frame change rates
-let animationIdleCount = 0;
-let previousChangeFrequency = 0; // Track previous change frequency
-let preRenderStartTime = 0;
-let preRenderMaxDuration = 30000; // Increased to 30 seconds for longer animations
-// Image2texture variables
-let animationDetectionSensitivity = 0.85; // Increased from 0.5 to 0.85 for much stricter detection by default
-let animationEndTime = 0;
-let animationStartTime = 0;
-let animationStartDetected = false;
-let animationDetected = false;
-export let preRenderedFrames = [];
-export let animationDuration = 0; // Store detected animation duration
-let preRenderingInProgress = false;
-export let isAnimationFinite = false;
-
-// Update references to use internal functions instead of imports
-// These variables will be exported and should be used by preview-util.js
-export let isPreviewActive = true; // Exported for use in preview-util.js
-export let isPreviewAnimationPaused = false; // Exported for use in preview-util.js
-export let lastTextureUpdateTime = 0; // Exported for use in preview-util.js and threejs-util.js
-
-// TIMING STATE - This module owns all animation timing
-let animationPlaybackStartTime = 0; // When playback actually started
-let animationCaptureStartTime = 0;  // When capture started (for offset calculations)
-// Active playback values
-let playbackStartTime = 0;
-let isPlaybackActive = false;
 
 /**
  * Initialize playback timing - called when preview starts playing
@@ -166,402 +166,6 @@ export function startPlayback() {
 export function stopPlayback() {
     isPlaybackActive = false;
     console.log('Playback stopped');
-}
-
-/**
- * Start pre-rendering animation frames
- * @param {HTMLIFrameElement} iframe - The iframe containing the HTML content
- * @param {Function} callback - Function to call when pre-rendering is complete
- * @param {HTMLElement} progressBar - Optional progress bar element to update
- * @param {CustomTextureSettings} settings - Optional settings object for texture configuration
- * @param {THREE.Mesh} previewPlane - The mesh to apply textures to
- */
-export function startImage2TexturePreRendering(iframe, callback, progressBar = null, settings = null, previewPlane = null) {
-    if (!iframe) {
-        console.error('No iframe provided for pre-rendering');
-        if (callback) callback();
-        return;
-    }
-    
-    // Reset state
-    preRenderingInProgress = true;
-    preRenderedFrames = [];
-    
-    // Set the start time
-    const preRenderStartTime = Date.now();
-    
-    // Track progress metrics
-    let totalFramesEstimate = 120; // Initial estimate
-    let lastProgressUpdate = 0;
-    let progressUpdateInterval = 100; // Update progress every 100ms
-    let maxProgressBeforeFinalAnimation = 92; // Cap progress at this value until final animation
-    finalProgressAnimation = false;
-    finalProgressStartTime = 0;
-    
-    // Track animation detection variables
-    let loopDetected = false;
-    let endDetected = false;
-    let analysisMetrics = {};
-    
-    // Get animation settings from passed settings object instead of DOM
-    let isLongExposureMode = false;
-    let playbackSpeed = 1.0;
-    
-    if (settings) {
-        // Use settings parameters instead of DOM elements
-        isLongExposureMode = settings.isLongExposureMode;
-        playbackSpeed = settings.playbackSpeed || 1.0;
-    } else {
-        // Fallback to DOM access if settings not provided (for backward compatibility)
-        const animationTypeSelect = document.getElementById('html-animation-type');
-        isLongExposureMode = animationTypeSelect && animationTypeSelect.value === 'longExposure';
-    }
-    
-    // Set flag if we're capturing for long exposure
-    if (isLongExposureMode) {
-        setCapturingForLongExposure(true);
-        
-        // Temporarily disable borders during capture
-        const originalBorderSetting = window.showPreviewBorders;
-        window.showPreviewBorders = false;
-        console.log('Borders temporarily disabled for long exposure capture');
-        
-        // Store original setting to restore later
-        window._originalBorderSetting = originalBorderSetting;
-    }
-    
-    // Function to update progress bar
-    const updateProgress = (percent) => {
-        if (progressBar) {
-            // Ensure progress never exceeds maxProgressBeforeFinalAnimation unless in final animation
-            if (!finalProgressAnimation && percent > maxProgressBeforeFinalAnimation) {
-                percent = maxProgressBeforeFinalAnimation;
-            }
-            progressBar.style.width = `${percent}%`;
-        }
-    };
-    
-    // Function to create the long exposure texture and apply it
-    const createAndApplyLongExposure = () => {
-        if (preRenderedFrames.length > 0) {
-            // Use the playbackSpeed from settings instead of DOM
-            const longExposureTexture = createLongExposureTexture(preRenderedFrames, playbackSpeed);
-            
-            // Update the mesh with the long exposure texture
-            if (previewPlane) {
-                updateMeshTexture(longExposureTexture, previewPlane);
-            }
-            
-            // Show a message about the long exposure
-            showStatus(`Long exposure created from ${preRenderedFrames.length} frames`, 'success');
-            
-            // Pause animation since we just want to display the static image
-            setIsPreviewAnimationPaused(true);
-        }
-    };
-    
-    // Function to start final progress animation
-    const startFinalProgressAnimation = () => {
-        if (finalProgressAnimation) return; // Already animating
-        
-        finalProgressAnimation = true;
-        finalProgressStartTime = Date.now();
-        
-        // Start the animation loop
-        animateFinalProgress();
-    };
-    
-    // Function to animate progress to 100%
-    const animateFinalProgress = () => {
-        const now = Date.now();
-        const elapsed = now - finalProgressStartTime;
-        
-        if (elapsed >= finalProgressDuration) {
-            // Animation complete, set to 100%
-            updateProgress(100);
-            
-            // Log animation analysis report
-            logAnimationAnalysisReport('Image2Texture', {
-                frameCount: preRenderedFrames.length,
-                duration: animationDuration,
-                isFinite: isAnimationFinite,
-                loopDetected,
-                endDetected,
-                analysisTime: now - preRenderStartTime,
-                metrics: analysisMetrics
-            });
-            
-            // Hide loading overlay with fade out
-            const loadingOverlay = document.getElementById('pre-rendering-overlay');
-            if (loadingOverlay) {
-                loadingOverlay.style.transition = 'opacity 0.5s ease';
-                loadingOverlay.style.opacity = '0';
-                
-                // Remove after fade out
-                setTimeout(() => {
-                    if (loadingOverlay.parentNode) {
-                        loadingOverlay.parentNode.removeChild(loadingOverlay);
-                    }
-                    
-                    // Now create the info panel after pre-rendering is complete
-                    const canvasContainer = document.querySelector('#html-preview-content');
-                    if (canvasContainer) {
-                        const modal = document.getElementById('html-editor-modal');
-                        const currentMeshId = parseInt(modal.dataset.meshId);
-                        createMeshInfoPanel(canvasContainer, currentMeshId);
-                    }
-                    
-                    // For long exposure, create the static image now that all frames are captured
-                    if (isLongExposureMode && preRenderedFrames.length > 0) {
-                        // Use playbackSpeed from settings instead of DOM
-                        const longExposureTexture = createLongExposureTexture(preRenderedFrames, playbackSpeed);
-                        
-                        // Update the mesh with the long exposure texture
-                        if (previewPlane) {
-                            updateMeshTexture(longExposureTexture, previewPlane);
-                        }
-                        
-                        // Show a message about the long exposure
-                        showStatus(`Long exposure created from ${preRenderedFrames.length} frames`, 'success');
-                        
-                        // Pause animation since we just want to display the static image
-                        setIsPreviewAnimationPaused(true);
-                    } else {
-                        // Reset animation start time to now
-                        startPlayback();
-                        
-                        // Start the animation
-                        setIsPreviewAnimationPaused(false);
-                        
-                        // Show a message that playback is starting
-                        showStatus(`Animation playback starting at ${playbackSpeed}x speed`, 'success');
-                    }
-                }, 500);
-                
-                // Don't continue the animation
-                return;
-            }
-        } else {
-            // Calculate progress based on easing function
-            const progress = easeOutCubic(elapsed / finalProgressDuration);
-            const currentProgress = maxProgressBeforeFinalAnimation + (100 - maxProgressBeforeFinalAnimation) * progress;
-            updateProgress(currentProgress);
-            
-            // Update loading text
-            const progressText = document.getElementById('loading-progress-text');
-            if (progressText) {
-                if (isLongExposureMode) {
-                    progressText.textContent = 'Creating long exposure...';
-                } else {
-                    progressText.textContent = 'Finalizing animation...';
-                }
-            }
-            
-            // Continue animation
-            requestAnimationFrame(animateFinalProgress);
-        }
-    };
-    
-    // Easing function for smooth animation
-    const easeOutCubic = (x) => {
-        return 1 - Math.pow(1 - x, 3);
-    };
-    
-    // Create high-quality texture from iframe for better visuals
-    const createHighQualityTexture = async (iframe) => {
-        try {
-            const texture = await createTextureFromIframe(iframe);
-            
-            // Apply higher quality settings
-            texture.anisotropy = 16; // Increased from 8 for sharper textures
-            texture.minFilter = THREE.LinearFilter;
-            texture.magFilter = THREE.LinearFilter;
-            texture.generateMipmaps = false;
-            texture.needsUpdate = true;
-            
-            return texture;
-        } catch (error) {
-            console.error('Error creating high-quality texture:', error);
-            throw error;
-        }
-    };
-    
-    // Function to capture frames until animation completes or times out
-    const captureFrames = async () => {
-        if (!isPreviewActive || !preRenderingInProgress) {
-            preRenderingInProgress = false;
-            startFinalProgressAnimation();
-            return;
-        }
-        
-        const now = Date.now();
-        
-        // Update progress based on more accurate metrics
-        if (now - lastProgressUpdate > progressUpdateInterval) {
-            lastProgressUpdate = now;
-            
-            // Calculate elapsed time percentage
-            const elapsedTime = now - preRenderStartTime;
-            const timeProgress = Math.min(90, (elapsedTime / preRenderMaxDuration) * 100);
-            
-            // Calculate frame-based progress
-            let frameProgress = 0;
-            if (animationDetected) {
-                // If we've detected animation, adjust the total frames estimate
-                if (preRenderedFrames.length > totalFramesEstimate * 0.5) {
-                    // If we've captured more than half our estimate, update the estimate
-                    totalFramesEstimate = Math.max(totalFramesEstimate, Math.ceil(preRenderedFrames.length * 1.2));
-                }
-                frameProgress = Math.min(90, (preRenderedFrames.length / totalFramesEstimate) * 100);
-            } else {
-                // If no animation detected, use time-based progress
-                frameProgress = timeProgress;
-            }
-            
-            // Use a weighted combination of time and frame progress
-            // Cap at maxProgressBeforeFinalAnimation to leave room for final animation
-            const combinedProgress = Math.min(
-                maxProgressBeforeFinalAnimation, 
-                (timeProgress * 0.3) + (frameProgress * 0.7)
-            );
-            updateProgress(combinedProgress);
-            
-            // Update the loading text to show more information
-            const progressText = document.getElementById('loading-progress-text');
-            if (progressText) {
-                progressText.textContent = `Pre-rendering animation... ${preRenderedFrames.length} frames captured`;
-            }
-        }
-        
-        // Check if we've detected a loop and have enough frames
-        // Use the improved detection logic with higher sensitivity for pre-rendering
-        const sensitivity = animationDetectionSensitivity + 0.1; // Slight boost during pre-rendering
-        const isLoopDetected = preRenderedFrames.length > 20 && 
-                             detectAnimationLoop(preRenderedFrames, sensitivity);
-                             
-        if (isLoopDetected && preRenderedFrames.length > 20) {
-            console.log('Animation loop detected after ' + preRenderedFrames.length + ' frames');
-            preRenderingInProgress = false;
-            isAnimationFinite = true;
-            animationDuration = preRenderedFrames[preRenderedFrames.length - 1].timestamp - preRenderedFrames[0].timestamp;
-            
-            // Update analysis metrics
-            loopDetected = true;
-            
-            // Show success message
-            if (isLongExposureMode) {
-                showStatus(`Animation loop detected, creating long exposure from ${preRenderedFrames.length} frames`, 'info');
-            } else {
-                showStatus(`Animation loop detected (${(animationDuration/1000).toFixed(1)}s), ${preRenderedFrames.length} frames captured`, 'success');
-            }
-            
-            // Start final progress animation instead of immediately calling callback
-            startFinalProgressAnimation();
-            return;
-        }
-        
-        // Check for animation end using the new analysis function
-        if (preRenderedFrames.length > 20) { // Reduced from 30
-            // Calculate the latest frame hash
-            const latestFrameHash = preRenderedFrames[preRenderedFrames.length - 1].hash;
-            
-            // Analyze frames to detect animation end with higher sensitivity
-            const analysisResult = analyzeAnimationFrames(
-                preRenderedFrames.slice(0, -1), // All frames except the latest
-                latestFrameHash,
-                sensitivity
-            );
-            
-            // Store analysis metrics
-            analysisMetrics = analysisResult.metrics;
-            
-            // If end detected and we have enough frames, stop pre-rendering
-            if (analysisResult.endDetected && preRenderedFrames.length > 20) {
-                console.log('Animation end detected during pre-rendering after ' + preRenderedFrames.length + ' frames');
-                console.log('Detection metrics:', analysisResult.metrics);
-                preRenderingInProgress = false;
-                isAnimationFinite = true;
-                animationDuration = preRenderedFrames[preRenderedFrames.length - 1].timestamp - preRenderedFrames[0].timestamp;
-                
-                // Update analysis metrics
-                endDetected = true;
-                
-                // Show success message
-                showStatus(`Animation end detected (${(animationDuration/1000).toFixed(1)}s), ${preRenderedFrames.length} frames captured`, 'success');
-                
-                // Start final progress animation instead of immediately calling callback
-                startFinalProgressAnimation();
-                return;
-            }
-        }
-        
-        // Check if we've exceeded the maximum pre-rendering time
-        if (now - preRenderStartTime > preRenderMaxDuration) {
-            console.log('Pre-rendering time limit reached after ' + preRenderMaxDuration + 'ms');
-            preRenderingInProgress = false;
-            
-            if (loopDetected) {
-                isAnimationFinite = true;
-                animationDuration = preRenderedFrames[preRenderedFrames.length - 1].timestamp - preRenderedFrames[0].timestamp;
-                console.log(`Animation loop detected, duration: ${animationDuration}ms, ${preRenderedFrames.length} frames captured`);
-                showStatus(`Animation loop detected (${(animationDuration/1000).toFixed(1)}s), ${preRenderedFrames.length} frames captured`, 'success');
-            } else if (animationStartDetected && animationEndDetected) {
-                isAnimationFinite = true;
-                animationDuration = animationEndTime - animationStartTime;
-                console.log(`Animation start/end detected, duration: ${animationDuration}ms, ${preRenderedFrames.length} frames captured`);
-                showStatus(`Animation start/end detected (${(animationDuration/1000).toFixed(1)}s), ${preRenderedFrames.length} frames captured`, 'success');
-                
-                // Update analysis metrics
-                endDetected = true;
-            } else {
-                console.log(`No animation loop detected, ${preRenderedFrames.length} frames captured`);
-                showStatus(`Animation appears infinite, ${preRenderedFrames.length} frames captured for playback`, 'info');
-            }
-            
-            // Start final progress animation instead of immediately calling callback
-            startFinalProgressAnimation();
-            return;
-        }
-        
-        try {
-            // Capture a high-quality frame
-            const texture = await createHighQualityTexture(iframe);
-            
-            // Calculate a hash of the texture to detect changes
-            const frameHash = calculateTextureHash(texture);
-            
-            // Add frame to pre-rendered frames
-            preRenderedFrames.push({
-                texture: texture,
-                timestamp: now,
-                hash: frameHash
-            });
-            
-            // For long exposure mode, if we have enough frames, create the texture immediately
-            // This prevents showing the first frame before the long exposure
-            if (isLongExposureMode && preRenderedFrames.length >= 15) {
-                // Create the long exposure immediately
-                createAndApplyLongExposure();
-            }
-            
-            // Use a shorter delay for more frequent frame capture to increase smoothness
-            setTimeout(() => {
-                requestAnimationFrame(captureFrames);
-            }, 5); // 5ms delay allows for more frames to be captured in the same time
-        } catch (error) {
-            console.error('Error during pre-rendering:', error);
-            preRenderingInProgress = false;
-            
-            // Start final progress animation instead of immediately calling callback
-            startFinalProgressAnimation();
-        }
-    };
-    
-    // Start capturing frames
-    captureFrames();
-    
-    // Store callback to be called after final animation completes
-    window._preRenderCallback = callback;
 }
 
 /**
@@ -1391,14 +995,6 @@ function calculateHashDifference(hash1, hash2) {
 }
 
 /**
- * Set is preRenderingInProgress
- * @param {boolean} incomingValue - The new value to set
- */
-export function setPreRenderingInProgress(incomingValue) {
-    preRenderingInProgress = incomingValue;
-}
-
-/**
  * Update the mesh texture with the given texture
  * @param {THREE.Texture} texture - The texture to apply to the mesh
  * @param {THREE.Mesh} previewPlane - The mesh to update with the texture
@@ -1453,37 +1049,6 @@ export function setLastTextureUpdateTime(incomingValue) {
     lastTextureUpdateTime = incomingValue;
 }
 
-/**
- * Get isPreviewActive state
- * @returns {boolean} Current isPreviewActive value
- */
-export function getIsPreviewActive() {
-    return isPreviewActive;
-}
-
-/**
- * Get isPreviewAnimationPaused state
- * @returns {boolean} Current isPreviewAnimationPaused value
- */
-export function getIsPreviewAnimationPaused() {
-    return isPreviewAnimationPaused;
-}
-
-/**
- * Get lastTextureUpdateTime value
- * @returns {number} Current lastTextureUpdateTime value
- */
-export function getLastTextureUpdateTime() {
-    return lastTextureUpdateTime;
-}
-
-/**
- * Get preRenderedFrames array
- * @returns {Array} Current preRenderedFrames array
- */
-export function getPreRenderedFrames() {
-    return preRenderedFrames;
-}
 
 /**
  * Set preRenderedFrames array
@@ -1491,22 +1056,6 @@ export function getPreRenderedFrames() {
  */
 export function setPreRenderedFrames(incomingValue) {
     preRenderedFrames = incomingValue;
-}
-
-/**
- * Get preRenderingInProgress state
- * @returns {boolean} Current preRenderingInProgress value
- */
-export function getPreRenderingInProgress() {
-    return preRenderingInProgress;
-}
-
-/**
- * Get animationDuration value
- * @returns {number} Current animationDuration value
- */
-export function getAnimationDuration() {
-    return animationDuration;
 }
 
 /**
@@ -1518,14 +1067,6 @@ export function setAnimationDuration(incomingValue) {
 }
 
 /**
- * Get isAnimationFinite state
- * @returns {boolean} Current isAnimationFinite value
- */
-export function getIsAnimationFinite() {
-    return isAnimationFinite;
-}
-
-/**
  * Set isAnimationFinite state
  * @param {boolean} incomingValue - The new value to set
  */
@@ -1533,40 +1074,14 @@ export function setIsAnimationFinite(incomingValue) {
     isAnimationFinite = incomingValue;
 }
 
-/**
- * Calculate a simple hash of a texture to detect changes between frames
- * @param {THREE.Texture} texture - The texture to hash
- * @returns {string} A simple hash of the texture
- */
-function calculateTextureHash(texture) {
-    if (!texture || !texture.image) return '';
-    
-    try {
-        // Create a small canvas to sample the texture
-        const canvas = document.createElement('canvas');
-        const size = 16; // Small sample size for performance
-        canvas.width = size;
-        canvas.height = size;
-        const ctx = canvas.getContext('2d');
-        
-        // Draw the texture to the canvas
-        ctx.drawImage(texture.image, 0, 0, size, size);
-        
-        // Get image data
-        const imageData = ctx.getImageData(0, 0, size, size).data;
-        
-        // Sample pixels at regular intervals
-        const samples = [];
-        const step = 4 * 4; // Sample every 4th pixel (RGBA)
-        for (let i = 0; i < imageData.length; i += step) {
-            // Use just the RGB values (skip alpha)
-            samples.push(imageData[i], imageData[i+1], imageData[i+2]);
-        }
-        
-        // Create a simple hash from the samples
-        return samples.join(',');
-    } catch (e) {
-        console.error('Error calculating texture hash:', e);
-        return '';
-    }
+export function setPreRenderingInProgress(incomingValue) {
+    preRenderingInProgress = incomingValue;
+}
+
+export function setFinalProgressAnimation(incomingValue) {
+    finalProgressAnimation = incomingValue;
+}
+
+export function setFinalProgressStartTime(incomingValue) {
+    finalProgressStartTime = incomingValue;
 }
